Client-side & HTTP API authentication (#1)
* client side auth0 - init

* Add server w/ JWT authentication. Auth needs cleanup.

Note: @nuxtjs/auth@4.9.x is NOT stable. Versions since 2018 require 
nanoid to be installed as a direct dependency. Versions before this do 
not properly send 'audience' parameter to auth0 server, so the bearer 
token for API is not JWT as required (instead is opaque token)

* mv server env info to .env. NO SECRETS

cp .env.dev to .env for dev env config